# swagger.yaml
openapi: 3.0.0
info:
  title: Phone Store API
  version: 1.0.0
  description: Phone products REST API
servers:
  - url: http://localhost:3000
paths:
  /api/v1/phones:
    get:
      summary: Get all phones
      tags: [Phones]
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search by title
        - in: query
          name: brand
          schema:
            type: string
          description: Filter by brand
        - in: query
          name: minPrice
          schema:
            type: number
          description: Minimum price
        - in: query
          name: maxPrice
          schema:
            type: number
          description: Maximum price
        - in: query
          name: page
          schema:
            type: number
          description: Page number
        - in: query
          name: limit
          schema:
            type: number
          description: Items per page
      responses:
        "200":
          description: Success
        "400":
          description: Failed
    post:
      summary: Create new phone (for Admin User)
      tags: [Phones]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brand:
                  type: string
                model:
                  type: string
                title:
                  type: string
                description:
                  type: string
                price:
                  type: object
                  properties:
                    current:
                      type: number
                    oldPrice:
                      type: number
                stock:
                  type: number
                color:
                  type: string
                images:
                  type: array
                  items:
                    type: string
                colors:
                  type: array
                  items:
                    type: string
                ram:
                  type: string
                processor:
                  type: string
                chipset:
                  type: string
                screenType:
                  type: string
      responses:
        "201":
          description: Created
        "400":
          description: Failed
...
  /api/v1/cart:
    get:
      summary: Get Your Cart
      tags: [Cart]
      responses:
        "201":
          description: Created
        "400":
          description: Failed
    post:
      summary: Add Product To Cart
      tags: [Cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                quantity:
                  type: string      
      responses:
        "201":
          description: Created
        "400":
          description: Failed                      
    patch:
      summary: Update Product In Cart
      tags: [Cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                quantity:
                  type: string      
      responses:
        "201":
          description: Created
        "400":
          description: Failed 
...